Fix some mypy issues:
  * Re-type sleep to float (as it seems that is actually the way it is used)
  * Add some typing to funcs

+++ test/base/env.py
@@ -15,7 +15,7 @@ class MyTestEnv(gym.Env):
     def __init__(
         self,
         size: int,
-        sleep: int = 0,
+        sleep: float = 0,
         dict_state: bool = False,
         recurse_state: bool = False,
         ma_rew: int = 0,
